+++ app/src/androidTest/java/es/uc3m/android/bloom/AndroidTest.java
@@ -14,16 +14,20 @@
  * limitations under the License.
  *
  */
-package es.uc3m.android.navigation;
+package es.uc3m.android.bloom;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
- * Example local unit test, which will execute on the development machine (host).
+ * Instrumented test, which will execute on an Android device.
  *
  * @see <a href="https://developer.android.com/studio/test">Testing documentation</a>
  */
-public class UnitTest {
+@RunWith(AndroidJUnit4.class)
+public class AndroidTest {
 
     @Test
     public void test() {
@@ -31,5 +35,4 @@ public class UnitTest {
 
         // TODO: Verify
     }
-
 }
