+++ app/src/main/java/es/uc3m/android/bloom/ProfileFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  *
  */
-package es.uc3m.android.navigation;
+package es.uc3m.android.bloom;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -23,12 +23,12 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-public class HomeFragment extends Fragment {
+public class ProfileFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.home_fragment, container, false);
+        return inflater.inflate(R.layout.profile_fragment, container, false);
     }
 
 }
