+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="home_fragment">Home</string>
     <string name="profile_fragment">Profile</string>
     <string name="settings_fragment">Settings</string>
+    <string name="register_button">Register</string>
 </resources>
